Override global record with new sourceVersion
This means there is only one sourceVersion for a particular cluster.
So output is latest always (if that global record is updated).
Global records not part of new version can be deleted as usual.
Since the deletion only touches those with new version the OAI
datestamp will be optimal there.