Make up default leader for Goldrush
This means that no exception is thrown if leader is absent. If leader
is absent, it's also absent in payload. So one can determine which records
have missing leader (after they have been ingested).
Goldrush key only uses leader pos 6: material type which defaults
to 'a' with the default leader.